[Mod Manager] - Move Panacea, Backend, and SteamAPI stuff to a service.
